requirements.txt
pytubefix
customtkinter
pillow